add middleware support in `route(...)` calls
This change allows passing in of middleware functions in your `route()`
calls. Middleware functions will receive a `callable $next` as its first
argument when invoked, followed by `array $params`, and the rest of the
arguments that were passed to `dispatch()`. Note that the `$params`
argument will always be there, compared to its optional nature when
actions are called.